+++ config/butler.php
@@ -21,6 +21,8 @@ return [
 
         'schema' => env('BUTLER_GRAPHQL_SCHEMA', app_path('Http/Graphql/schema.graphql')),
 
+        'schema_extensions_path' => env('BUTLER_GRAPHQL_SCHEMA_EXTENSIONS_PATH', app_path('Http/Graphql/')),
+        'schema_extensions_glob' => env('BUTLER_GRAPHQL_SCHEMA_EXTENSIONS_GLOB', 'schema-*.graphql'),
     ],
 
     'health' => [