+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Dropdown component for navbar.
 - Test method `assertGraphqlSent`.
+- butler-graphql schema extension config.
 
 ## [0.24.2] - 2023-11-15