+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.25.0] - 2024-02-05
+
 ### Added
 - Dropdown component for navbar.
 - Test method `assertGraphqlSent`.